requirements.txt
Flask
python-doctr
pdf2image
Pillow
doctr
tf2onnx
tensorflow
